requirements-dev.txt
# SPDX-License-Identifier: Apache-2.0

-r requirements.txt

llama-cpp-python[server]>=0.3.0,<1.0.0
pre-commit>=3.0.4,<5.0
pylint>=2.16.2,<4.0
pylint-pydantic
pytest
pytest-asyncio
pytest-cov
pytest-html
tox>=4.4.2,<5
jsonschema
